[H] NetworkMap 클래스 생성 및 함수 이동, 기존 Network클래스 수정
