+++ Project/MachumCafe/MachumCafe/ListMapViewController.swift
@@ -9,8 +9,6 @@
 import UIKit
 import GoogleMaps
 import GooglePlaces
-import Alamofire
-import SwiftyJSON
 
 class ListMapViewController: UIViewController,GMSMapViewDelegate,CLLocationManagerDelegate {
 
@@ -21,7 +19,6 @@ class ListMapViewController: UIViewController,GMSMapViewDelegate,CLLocationManag
     var locationManager = CLLocationManager()
     var placesClient: GMSPlacesClient!
     var currentLocation = [String: Double]()
-    let urlMapKey = "AIzaSyBNTjHJ-wYRN_p9x7HMJu-_sI2LG-kzVj4"
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -49,11 +46,12 @@ class ListMapViewController: UIViewController,GMSMapViewDelegate,CLLocationManag
     func mapView(_ mapView: GMSMapView, didTapAt coordinate: CLLocationCoordinate2D) {
         googleMap.clear()
         currentLocation["latitude"] = coordinate.latitude
-        currentLocation["longtitude"] = coordinate.longitude
+        currentLocation["longitude"] = coordinate.longitude
         
-        
-        searchCafeAroundMe(latitude: coordinate.latitude, longtitude: coordinate.longitude) { (cafeInfoArray) in
-            self.spreadMarker(cafeInfoArray: cafeInfoArray)
+        NetworkMap.searchCafeAroundMe(latitude: coordinate.latitude, longitude: coordinate.longitude) { (cafeInfoArray) in
+            NetworkMap.spreadMarker(cafeInfoArray: cafeInfoArray, callback: { (cafeInfo, name, address) in
+                self.createMarker(titleMarker: name, snippetMarker: address, iconMarker: #imageLiteral(resourceName: "marker") , latitude: cafeInfo["cafeLatitude"]! as!CLLocationDegrees, longitude: cafeInfo["cafeLongitude"]! as! CLLocationDegrees)
+            })
         }
         
         createMarker(titleMarker: "", snippetMarker: "", iconMarker: #imageLiteral(resourceName: "locationIcon"), latitude: coordinate.latitude, longitude: coordinate.longitude)
@@ -74,49 +72,5 @@ class ListMapViewController: UIViewController,GMSMapViewDelegate,CLLocationManag
         marker.icon = iconMarker
         marker.map = googleMap
     }
-    
-    // Part - Method : 내 주위 700미터 안의 카페 탐색
-    func searchCafeAroundMe (latitude : CLLocationDegrees, longtitude: CLLocationDegrees, completion : @escaping ([[String : Any]]) -> Void) {
-        
-        var cafeInfo = [String : Any]()
-        var cafeInfoArray = [[String : Any]]()
-        
-        let aroundMeURL = "https://maps.googleapis.com/maps/api/place/radarsearch/json?location=\(latitude),\(longtitude)&radius=700&type=cafe&key=\(urlMapKey)"
-        
-        Alamofire.request(aroundMeURL).responseJSON { responds in
-            let json = JSON(data: responds.data!)
-            let results = json["results"].arrayValue
-            
-            for result in results {
-                let geometry = result["geometry"].dictionaryValue
-                let location = geometry["location"]?.dictionary
-                let cafeLatitude = location?["lat"]?.doubleValue
-                let cafeLongtitude = location?["lng"]?.doubleValue
-                let cafeInfoID = result["place_id"].stringValue
-                
-                cafeInfo["cafeLatitude"] = cafeLatitude
-                cafeInfo["cafeLongtitude"] = cafeLongtitude
-                cafeInfo["cafeInfoID"] = cafeInfoID
-                cafeInfoArray.append(cafeInfo)
-            }
-            completion(cafeInfoArray)
-        }
-    }
-    // Part - Method : 마커 뿌리기
-    func spreadMarker (cafeInfoArray : [[String : Any]]) {
-        for cafeInfo in cafeInfoArray {
-            let cafeURL = "https://maps.googleapis.com/maps/api/place/details/json?placeid=\(String(describing: cafeInfo["cafeInfoID"]!))&language=ko&key=\(urlMapKey)"
-            
-            Alamofire.request(cafeURL).responseJSON { responds in
-                let json = JSON(data: responds.data!)
-                let result = json["result"].dictionary
-                let address = result?["formatted_address"]?.stringValue
-                let name = result?["name"]?.stringValue
-                
-                self.createMarker(titleMarker: name!, snippetMarker: address!, iconMarker: #imageLiteral(resourceName: "marker") , latitude: cafeInfo["cafeLatitude"]! as!CLLocationDegrees, longitude: cafeInfo["cafeLongtitude"]! as! CLLocationDegrees)
-            }
-        }
-    }
-
     
 }
