+++ Project/MachumCafe/MachumCafe/NetworkBookmark.swift
@@ -10,9 +10,11 @@ import Foundation
 import Alamofire
 
 class NetworkBookmark {
+    
+    private static let url = URLpath.getURL()
+    
     // MARK: 즐겨찾기 목록 데이터모델에 저장
     static func getMyBookmark(userId: String, callback: @escaping (_ message: Bool, _ cafeList: [ModelCafe], _ userBookmark: [String]) -> Void) {
-        let url = URLpath.getURL()
         var message = Bool()
         var cafeList = [ModelCafe]()
         var userBookmark = [String]()
@@ -57,7 +59,6 @@ class NetworkBookmark {
     
     // MARK: 카페 이미지 데이터모델에 저장
     static func getImagesData(imagesName: [String], cafe: ModelCafe) {
-        let url = URLpath.getURL()
         
         for imageName in imagesName {
             Alamofire.request("\(url)/api/v1/cafe/\(imageName)").responseData(completionHandler: { (response) in
@@ -68,7 +69,6 @@ class NetworkBookmark {
     
     // MARK: 즐겨찾기 추가 & 삭제
     static func setMyBookmark(userId: String, cafeId: String, callback: @escaping (_ message: Bool, _ description: String) -> Void) {
-        let url = URLpath.getURL()
         var message = Bool()
         var description = String()
